keep buffer-like temp file in memory
When searching the header or body for strings and the `thorough_search'
option is set, a temp file was created, parsed, and then unlinked again.
This is now done in memory using glibc's open_memstream() and
fmemopen() if they are available.

This is a bit hackish, but might increase performance greatly. YMMV.

Signed-off-by: Julius Plenz <plenz@cis.fu-berlin.de>